requirements.txt
flask
supabase
requests
pandas
plotly
python-dotenv
transformers
torch
gunicorn
accelerate
